Refactoring: renamed structs, enums, functions in cgw_ucentral_parser.rs
